fixed indentation, added error handling and comments, removed captcha functionality

+++ client/src/app/login/login.component.ts
@@ -1,10 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { AbstractControl, FormGroup, FormControl, Validators } from '@angular/forms';
-import { Router, NavigationExtras } from '@angular/router';
-import { LoginData } from '../models/loginData';
+import { FormGroup, FormControl, Validators } from '@angular/forms';
+import { Router } from '@angular/router';
 import { AuthenticationService } from '../services/authentication.service';
-import { ToastrService } from '../services/toastr.service';
-
+import { ToastrService } from '../services/toastr.service'; //used to show error or success messages
 
 @Component({
   selector: 'app-login',
@@ -13,8 +11,9 @@ import { ToastrService } from '../services/toastr.service';
 })
 export class LoginComponent implements OnInit {
      
-  loginForm: FormGroup;
-  isBusy: boolean = false;
+  loginForm: FormGroup; //used to caputure username and password
+  isBusy: boolean = false; //prevent multiple submits of form, once user sends request becomes true until server response.
+
   constructor(private authenticationService: AuthenticationService, private showMessage: ToastrService, private router: Router) { 
     this.loginForm = new FormGroup({
       email: new FormControl(null, Validators.email),
@@ -27,27 +26,30 @@ export class LoginComponent implements OnInit {
   }
 
   checkIfLoggedIn() {
-    
-    this.authenticationService.checkIfLoggedIn().subscribe ( response => {
-      this.router.navigate(['/dashboard']);
-    }, 
-    err => {  
-     if(err.status == 400) {
-      //The token though present in local storage has expired or not present. 
-      //clear localStorage here. 
-      localStorage.clear();
-     }
-    });
+    if(localStorage.getItem('token') != null) {
+      this.authenticationService.checkIfLoggedIn().subscribe (response => {
+        this.router.navigate(['/dashboard']);
+      }, err => {  
+      if(err.status == 400) {
+        //The token though present in local storage has expired or not present. 
+        //clear localStorage here. 
+        localStorage.clear();
+      }
+      });
+    }
   }
 
+  //this is used to check if the form is valid by passing each field's name from html
   isValid(controlName) {
     return this.loginForm.get(controlName).invalid && this.loginForm.get(controlName).touched;
   }
 
+  //this method would call server to validate the credentials
   validateLogin() {
     this.isBusy = true;
     if(this.loginForm.valid) {
       this.authenticationService.validateLogin(this.loginForm.value).subscribe(response => {
+        //if login is successful, store data in local storage to be used for token and access.
         localStorage.setItem('token', response.token);
         localStorage.setItem('userID', response.userID);
         localStorage.setItem('firstName', response.firstName);
@@ -60,7 +62,5 @@ export class LoginComponent implements OnInit {
         this.showMessage.showError(err.error);
       }); 
     }
- 
   }
-
 }
